error fix als klusjesman geen rating heeft. Rating als float opgehaald en verwerkt. Aantal bieders weergegeven op elke pagina en ook op info pagina aangepast. maincontroller profiel pagina aangepast zodat klusjesmannen juiste tasks zien. En nog veel meer, zie logboek :)

+++ Project/src/main/java/com/AppArch/Project/Service/TaskRepoService.java
@@ -18,14 +18,14 @@ public interface TaskRepoService {
 	public List<Task> getUserTasksDone(User u);
 	public List<Task> findOpenTasks();
 	public List<Task> findClosedTasks();
-	public int getAverageRating(User executor);
+	public float getAverageRating(User executor);
 	public void addTask(Task t);
 	public Optional<Task> getTaskById(int id);
 	public void updateTask(int id, String title, String description, float price);
 	public void remove(int id);
 	public void changeState(int id, State s);
 	public void addExecutor(int id, User executor);
-	public void reviewTask(int id, int rating);
+	public void reviewTask(int id, float rating);
 	public void updateStatus(int id, State status);
 	
 }
